Crude hack to force in context from a file.
I believe doing this is necessary because the `b:workspace_folder` and
`g:copilot_workspace_folders` appear to have no effect for me, at least
with vim version 9.1. I have not investigated super deeply, but I'm
guessing files from those directories are only used by the language
server, rather than being added wholesale to the transmitted context.
This is annoying for obvious reasons when it comes to files that the
language server won't do anything with, in my case SQL.